feat: Use the native color picker if supported
This presents a slight UX challenge in that the native color picker is uncontrollable from javascript. There is no way to close it or know that it has been closed. Therefore I've dealt with this by 'closing' the aha color picker when the native one opens and handing it control.